Refactor index management client integration tests for scalability (#67917)
