+++ x-pack/plugins/index_management/__jest__/client_integration/helpers/index.ts
@@ -4,18 +4,44 @@
  * you may not use this file except in compliance with the Elastic License.
  */
 
-import { setup as homeSetup } from './home.helpers';
-import { setup as templateCreateSetup } from './template_create.helpers';
-import { setup as templateCloneSetup } from './template_clone.helpers';
-import { setup as templateEditSetup } from './template_edit.helpers';
-
 export { nextTick, getRandomString, findTestSubject, TestBed } from '../../../../../test_utils';
 
-export { setupEnvironment } from './setup_environment';
+export { setupEnvironment, WithAppDependencies, services } from './setup_environment';
 
-export const pageHelpers = {
-  home: { setup: homeSetup },
-  templateCreate: { setup: templateCreateSetup },
-  templateClone: { setup: templateCloneSetup },
-  templateEdit: { setup: templateEditSetup },
-};
+export type TestSubjects =
+  | 'aliasesTab'
+  | 'appTitle'
+  | 'cell'
+  | 'closeDetailsButton'
+  | 'createTemplateButton'
+  | 'deleteSystemTemplateCallOut'
+  | 'deleteTemplateButton'
+  | 'deleteTemplatesConfirmation'
+  | 'documentationLink'
+  | 'emptyPrompt'
+  | 'manageTemplateButton'
+  | 'mappingsTab'
+  | 'noAliasesCallout'
+  | 'noMappingsCallout'
+  | 'noSettingsCallout'
+  | 'indicesList'
+  | 'indicesTab'
+  | 'indexTableIncludeHiddenIndicesToggle'
+  | 'indexTableIndexNameLink'
+  | 'reloadButton'
+  | 'reloadIndicesButton'
+  | 'row'
+  | 'sectionError'
+  | 'sectionLoading'
+  | 'settingsTab'
+  | 'summaryTab'
+  | 'summaryTitle'
+  | 'systemTemplatesSwitch'
+  | 'templateDetails'
+  | 'templateDetails.manageTemplateButton'
+  | 'templateDetails.sectionLoading'
+  | 'templateDetails.tab'
+  | 'templateDetails.title'
+  | 'templateList'
+  | 'templateTable'
+  | 'templatesTab';
