+++ x-pack/plugins/index_management/__jest__/client_integration/helpers/setup_environment.tsx
@@ -3,6 +3,7 @@
  * or more contributor license agreements. Licensed under the Elastic License;
  * you may not use this file except in compliance with the Elastic License.
  */
+
 /* eslint-disable @kbn/eslint/no-restricted-paths */
 import React from 'react';
 import axios from 'axios';
