+++ x-pack/plugins/index_management/__jest__/client_integration/index_template_wizard/constants.ts
