+++ x-pack/plugins/index_management/__jest__/client_integration/index_template_wizard/template_clone.helpers.ts
@@ -5,16 +5,16 @@
  */
 
 import { registerTestBed, TestBedConfig } from '../../../../../test_utils';
-import { BASE_PATH } from '../../../common/constants';
 import { TemplateClone } from '../../../public/application/sections/template_clone'; // eslint-disable-line @kbn/eslint/no-restricted-paths
+import { WithAppDependencies } from '../helpers';
+
 import { formSetup } from './template_form.helpers';
 import { TEMPLATE_NAME } from './constants';
-import { WithAppDependencies } from './setup_environment';
 
 const testBedConfig: TestBedConfig = {
   memoryRouter: {
-    initialEntries: [`${BASE_PATH}clone_template/${TEMPLATE_NAME}`],
-    componentRoutePath: `${BASE_PATH}clone_template/:name`,
+    initialEntries: [`/clone_template/${TEMPLATE_NAME}`],
+    componentRoutePath: `/clone_template/:name`,
   },
   doMountAsync: true,
 };
