--- x-pack/plugins/index_management/__mocks__/ui/notify.js
@@ -1,15 +1,0 @@
-/*
- * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
- * or more contributor license agreements. Licensed under the Elastic License;
- * you may not use this file except in compliance with the Elastic License.
- */
-
-export const toastNotifications = {
-  addInfo: () => {},
-  addSuccess: () => {},
-  addDanger: () => {},
-  addWarning: () => {},
-  addError: () => {},
-};
-
-export function fatalError() {}
