+++ x-pack/plugins/index_management/public/application/sections/home/index_list/index.ts
@@ -4,4 +4,4 @@
  * you may not use this file except in compliance with the Elastic License.
  */
 
-export * from './index_list';
+export { IndexList } from './index_list';
